+++ src/clustering/mod.rs
@@ -10,6 +10,7 @@ pub(crate) mod graph;
 pub mod hdbscan;
 pub mod kmeans;
 pub mod kprototypes;
+pub mod nmf;
 pub(crate) mod rng;
 pub mod scalespace;
 pub mod spectral;